More on signIn , logIn and logout
creating more logic to make more efficient our authentication functions. I did the styling on the navbar then i did the some authentication logic to choose the signIn or logIn for the user and when they choose one of them then they will only get the logout option. Also to do any task on the website user need to do login or signout.